initial restructuring of client code
